Added functionality to the create and edit event forms
Added functionality to the create and edit event forms and will now store data entered into it into the database. I also edited some values of the database so that it can store the correct data types to prevent a user from entering impossible values (e.g. half a ticket)

Date checking still needs to be added and the event poster (current user) functionality still needs to be included