Completed event database and altered login and CreateEdit forms
Completed the database functionality of the Events database table and now includes the id of the person who posted the event (this could be changed to the name just as easily but for now id works)

Changed the log in to field to a password field (was previously string).

Made sure to communicate to the user that we require 24hr time for our event fields.